requirements.txt
CairoSVG==2.5.2
Pillow==8.4.0
PyYAML==6.0
requests==2.21.0
spidev==3.5
